requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.1.2
django-ckeditor==6.7.1
django-js-asset==2.2.0
google==3.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.20.0
google-api-python-client==2.148.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
httplib2==0.22.0
idna==3.10
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.1.4
PyPDF2==3.0.1
python-decouple==3.8
requests==2.32.3
rsa==4.9
soupsieve==2.6
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
